Rename list to models context Fixes #25
