+++ packages/cms/src/pages/system/Models/Fields.js
@@ -2,7 +2,7 @@ import React from "react";
 import { CCard, CCardBody, CCardHeader } from "@coreui/react";
 import CIcon from "@coreui/icons-react";
 import { Link } from "react-router-dom";
-import { useList } from "context/list";
+import { useList } from "context/models";
 import { useModal } from "context/modal";
 import Info from "./Info";
 import FieldList from "./FieldList";
