+++ packages/cms/src/pages/system/Models/List.js
@@ -11,7 +11,7 @@ import {
 import CIcon from "@coreui/icons-react";
 import { useHistory, useParams } from "react-router-dom";
 import Spinner from "components/Spinner";
-import { useList } from "context/list";
+import { useList } from "context/models";
 import { useModal } from "context/modal";
 
 const Models = () => {
@@ -19,10 +19,6 @@ const Models = () => {
   const history = useHistory();
   const { dispatch, list, selected, hydrating } = useList();
   const { setModal } = useModal();
-
-  useEffect(() => {
-    dispatch({ type: "INIT", payload: "Model" });
-  }, [dispatch]);
 
   useEffect(() => {
     if (id && selected && id === selected) {
