+++ packages/cms/src/pages/system/Models/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 
-import { ListProvider } from "context/list";
 import { ModalProvider } from "context/modal";
 import { SelectProvider } from "context/select";
 import Models from "./Models";
@@ -9,13 +8,11 @@ import "scss/components/models.scss";
 
 const Index = () => {
   return (
-    <ListProvider>
-      <ModalProvider>
-        <SelectProvider>
-          <Models />
-        </SelectProvider>
-      </ModalProvider>
-    </ListProvider>
+    <ModalProvider>
+      <SelectProvider>
+        <Models />
+      </SelectProvider>
+    </ModalProvider>
   );
 };
